+++ src/rooms/rooms.controller.ts
@@ -10,7 +10,8 @@ import {
 	HttpStatus,
 	NotFoundException,
 	CacheInterceptor,
-	UseInterceptors
+	UseInterceptors,
+	ForbiddenException
 } from '@nestjs/common';
 import {
 	ApiBody,
@@ -25,6 +26,7 @@ import { AuthService } from '@/auth/auth.service';
 import { CreateRoomDto, RoomUserDto, UpdateRoomDto } from './dto';
 import { AuthToken } from '@/decorators/auth-token.decorator';
 import { Auth } from '@/decorators/auth.decorator';
+import { SecurityUserDto } from '@/users/dto';
 
 @ApiTags('Комнаты')
 @Controller('rooms')
@@ -41,6 +43,7 @@ export class RoomsController {
 		status: HttpStatus.OK,
 		type: Room,
 		isArray: true,
+		description: 'Все комнаты, в которых состоит пользователь',
 	})
 	@Auth()
 	@UseInterceptors(CacheInterceptor)
@@ -71,6 +74,20 @@ export class RoomsController {
 		return this.roomsService.getOne(id);
 	}
 
+	@ApiOperation({
+		summary: 'Получение всех пользователей комнаты',
+	})
+	@ApiResponse({
+		status: HttpStatus.OK,
+		type: SecurityUserDto,
+		isArray: true,
+		description: 'Пользователи комнаты',
+	})
+	@ApiResponse({
+		status: HttpStatus.NOT_FOUND,
+		type: NotFoundException,
+		description: 'Такой комнаты не существует',
+	})
 	@UseInterceptors(CacheInterceptor)
 	@Get('/:id/users')
 	async getUsers(@Param('id', ParseIntPipe) id: number) {
@@ -112,29 +129,78 @@ export class RoomsController {
 	@ApiResponse({
 		status: HttpStatus.OK,
 		type: Room,
+		description: 'Обновленная комната',
 	})
 	@ApiResponse({
 		status: HttpStatus.NOT_FOUND,
 		type: NotFoundException,
+		description: 'Такой комнаты не существует',
+	})
+	@ApiResponse({
+		status: HttpStatus.FORBIDDEN,
+		type: ForbiddenException,
+		description: 'Пользователь не может совершать действия в данной комнате',
 	})
 	@Auth()
 	@Put('/:id/update')
 	async update(
 		@Param('id', ParseIntPipe) id: number,
+		@AuthToken() token: string,
 		@Body() dto: UpdateRoomDto
 	): Promise<Room> {
+		const { id: userId, } = await this.authService.verifyUser(token);
+		const roomExistsUser = await this.roomsService.roomExistsUser(id, userId);
+		if (!roomExistsUser) {
+			throw new ForbiddenException('You dont have access');
+		}
 		return this.roomsService.update(id, dto);
 	}
 
+	@ApiOperation({
+		summary: 'Добавление пользователя в комнату',
+	})
+	@ApiBody({
+		description: 'Id пользователя для добавления',
+		type: RoomUserDto,
+	})
+	@ApiResponse({
+		status: HttpStatus.OK,
+		type: SecurityUserDto,
+		description: 'Добавленный пользователь',
+	})
+	@ApiResponse({
+		status: HttpStatus.FORBIDDEN,
+		type: ForbiddenException,
+		description: 'Пользователь не может совершать действия в данной комнате',
+	})
 	@Auth()
 	@Put('/:id/add-user')
 	async addUser(
 		@Param('id', ParseIntPipe) id: number,
+		@AuthToken() token: string,
 		@Body() dto: RoomUserDto
-	) {
+	): Promise<SecurityUserDto> {
+		const { id: userId, } = await this.authService.verifyUser(token);
+		const roomExistsUser = await this.roomsService.roomExistsUser(id, userId);
+		if (!roomExistsUser) {
+			throw new ForbiddenException('You dont have access');
+		}
 		return this.roomsService.addUser(id, dto);
 	}
 
+	@ApiOperation({
+		summary: 'Выход текущего пользователя из комнаты',
+	})
+	@ApiResponse({
+		status: HttpStatus.OK,
+		type: Boolean,
+		description: 'Удалось ли выйти',
+	})
+	@ApiResponse({
+		status: HttpStatus.FORBIDDEN,
+		type: ForbiddenException,
+		description: 'Пользователь не может совершать действия в данной комнате',
+	})
 	@Auth()
 	@Put('/:id/exit')
 	async removeUser(
@@ -142,6 +208,10 @@ export class RoomsController {
 		@AuthToken() token: string
 	) {
 		const { id: userId, } = await this.authService.verifyUser(token);
+		const roomExistsUser = await this.roomsService.roomExistsUser(id, userId);
+		if (!roomExistsUser) {
+			throw new ForbiddenException('You dont have access');
+		}
 		return this.roomsService.removeUser(id, { userId, });
 	}
 
@@ -150,11 +220,25 @@ export class RoomsController {
 	})
 	@ApiResponse({
 		status: HttpStatus.OK,
-		type: undefined,
+		type: Boolean,
+		description: 'Удалось ли выполнить удаление комнаты',
+	})
+	@ApiResponse({
+		status: HttpStatus.FORBIDDEN,
+		type: ForbiddenException,
+		description: 'Пользователь не может совершать действия в данной комнате',
 	})
 	@Auth()
 	@Delete('/:id/remove')
-	async remove(@Param('id', ParseIntPipe) id: number): Promise<boolean> {
+	async remove(
+		@Param('id', ParseIntPipe) id: number,
+		@AuthToken() token: string
+	): Promise<boolean> {
+		const { id: userId, } = await this.authService.verifyUser(token);
+		const roomExistsUser = await this.roomsService.roomExistsUser(id, userId);
+		if (!roomExistsUser) {
+			throw new ForbiddenException('You dont have access');
+		}
 		return this.roomsService.remove(id);
 	}
 }
